+++ src/code42cli/util.py
@@ -59,9 +59,20 @@ def print_no_existing_profile_message():
 
 
 def print_create_profile_help():
-    print(u"")
-    print(u"To add a profile, use: ")
-    print_bold(u"\tcode42 profile create <profile-name> <authority-URL> <username>")
+    print(u"\nTo add a profile, use: ")
+    print_bold(u"\tcode42 profile create <profile-name> <authority-URL> <username>\n")
+
+
+def print_set_default_profile_help(existing_profiles):
+    print(
+        u"\nNo default profile set.\n",
+        u"\nUse the --profile flag to specify which profile to use.\n",
+        u"\nTo set the default profile (used whenever --profile argument is not provided), use:"
+    )
+    print_bold(u"\tcode42 profile use <profile-name>")
+    print(u"\nExisting profiles:")
+    for profile in existing_profiles:
+        print("\t{}".format(profile))
     print(u"")